+++ test/src/date.test.js
@@ -4,6 +4,7 @@ import isInteger    from '../../src/number/isInteger';
 import isDate       from '../../src/date/is';
 import toUTC        from '../../src/date/toUTC';
 import startOfUTC   from '../../src/date/startOfUTC';
+import endOfUTC     from '../../src/date/endOfUTC';
 import addUTC       from '../../src/date/addUTC';
 import nowUnix      from '../../src/date/nowUnix';
 import nowUnixMs    from '../../src/date/nowUnixMs';
@@ -827,6 +828,213 @@ describe("Date", () => {
 
         it ('should return original date when passed a non-recognized key', () => {
             expect(startOfUTC(new Date("2023-05-01T12:04:27+02:00"), 'foobar')).to.eql(new Date("2023-05-01T12:04:27+02:00"));
+        });
+    });
+
+    describe("endOfUTC", () => {
+        it ('throw when passed a non-date', () => {
+            for (const el of [
+                ...fnStringValues(),
+                ...fnNumericValues(),
+                ...fnBooleanValues(),
+                ...fnRegexValues(),
+                ...fnObjectValues(),
+                ...fnNullables(),
+                ...fnArrayValues(),
+                ...fnFunctionValues(),
+            ]) {
+                expect(function () {
+                    endOfUTC(el, 'day');
+                }).to.throw('');
+            }
+        });
+
+        it ('throw when passed a non-string key', () => {
+            for (const el of [
+                ...fnDateValues(),
+                ...fnNumericValues(),
+                ...fnBooleanValues(),
+                ...fnRegexValues(),
+                ...fnObjectValues(),
+                ...fnNullables(),
+                ...fnArrayValues(),
+                ...fnFunctionValues(),
+            ]) {
+                expect(function () {
+                    endOfUTC(new Date(), el);
+                }).to.throw('');
+            }
+        });
+
+        it ('[year] should correctly set to end of year utc', () => {
+            expect(endOfUTC(new Date("2023-05-04T12:04:27+02:00"), 'year')).to.eql(new Date("2023-12-31T23:59:59.999Z"));
+        });
+
+        it ('[year] should be blazing fast in its conversion (1.000.000 benchmark < 750ms)', () => {
+            let start_time = getTime();
+            for (let i = 0; i < 1000000; i++) {
+                endOfUTC(new Date("2023-05-01T12:04:27+02:00"), 'year');
+            }
+            expect(getTime() - start_time).to.be.lt(750);
+        });
+
+        it ('[quarter] should correctly set to end of quarter utc', () => {
+            const qmap = {
+                1: {m: 3, d: 31},
+                2: {m: 3, d: 31},
+                3: {m: 3, d: 31},
+                4: {m: 6, d: 30},
+                5: {m: 6, d: 30},
+                6: {m: 6, d: 30},
+                7: {m: 9, d: 30},
+                8: {m: 9, d: 30},
+                9: {m: 9, d: 30},
+                10: {m: 12, d: 31},
+                11: {m: 12, d: 31},
+                12: {m: 12, d: 31},
+            };
+            for (let i = 1; i <= 12; i++) {
+                let date = `2023-${i < 10 ? '0' : ''}${i}-04T12:04:27+02:00`;
+                let date_q = `2023-${qmap[i].m < 10 ? '0' : ''}${qmap[i].m}-${qmap[i].d}T23:59:59.999Z`;
+                expect(endOfUTC(new Date(date), 'quarter')).to.eql(new Date(date_q));
+            }
+        });
+
+        it ('[quarter] should be blazing fast in its conversion (1.000.000 benchmark < 750ms)', () => {
+            let start_time = getTime();
+            for (let i = 0; i < 1000000; i++) {
+                endOfUTC(new Date("2023-05-01T12:04:27+02:00"), 'quarter');
+            }
+            expect(getTime() - start_time).to.be.lt(750);
+        });
+
+        it ('[month] should correctly set to end of month utc', () => {
+            expect(endOfUTC(new Date("2023-01-04T12:04:27+02:00"), 'month')).to.eql(new Date("2023-01-31T23:59:59.999Z"));
+            expect(endOfUTC(new Date("2023-02-04T12:04:27+02:00"), 'month')).to.eql(new Date("2023-02-28T23:59:59.999Z"));
+            expect(endOfUTC(new Date("2023-03-04T12:04:27+02:00"), 'month')).to.eql(new Date("2023-03-31T23:59:59.999Z"));
+            expect(endOfUTC(new Date("2023-04-04T12:04:27+02:00"), 'month')).to.eql(new Date("2023-04-30T23:59:59.999Z"));
+            expect(endOfUTC(new Date("2023-05-04T12:04:27+02:00"), 'month')).to.eql(new Date("2023-05-31T23:59:59.999Z"));
+            expect(endOfUTC(new Date("2023-06-04T12:04:27+02:00"), 'month')).to.eql(new Date("2023-06-30T23:59:59.999Z"));
+            expect(endOfUTC(new Date("2023-07-04T12:04:27+02:00"), 'month')).to.eql(new Date("2023-07-31T23:59:59.999Z"));
+            expect(endOfUTC(new Date("2023-08-04T12:04:27+02:00"), 'month')).to.eql(new Date("2023-08-31T23:59:59.999Z"));
+            expect(endOfUTC(new Date("2023-09-04T12:04:27+02:00"), 'month')).to.eql(new Date("2023-09-30T23:59:59.999Z"));
+            expect(endOfUTC(new Date("2023-10-04T12:04:27+02:00"), 'month')).to.eql(new Date("2023-10-31T23:59:59.999Z"));
+            expect(endOfUTC(new Date("2023-11-04T12:04:27+02:00"), 'month')).to.eql(new Date("2023-11-30T23:59:59.999Z"));
+            expect(endOfUTC(new Date("2023-12-04T12:04:27+02:00"), 'month')).to.eql(new Date("2023-12-31T23:59:59.999Z"));
+        });
+
+        it ('[month] should correctly set to end of month utc for february when in a leap year', () => {
+            expect(endOfUTC(new Date("2024-02-04T12:04:27+02:00"), 'month')).to.eql(new Date("2024-02-29T23:59:59.999Z"));
+        });
+
+        it ('[month] should be blazing fast in its conversion (1.000.000 benchmark < 750ms)', () => {
+            let start_time = getTime();
+            for (let i = 0; i < 1000000; i++) {
+                endOfUTC(new Date("2023-05-01T12:04:27+02:00"), 'month');
+            }
+            expect(getTime() - start_time).to.be.lt(750);
+        });
+
+        it ('[week] should correctly set to end of week utc with monday as first day of the week', () => {
+            expect(endOfUTC(new Date("2023-05-04T12:04:27+02:00"), 'week')).to.eql(new Date("2023-05-07T23:59:59.999Z"));
+            expect(endOfUTC(new Date("2023-05-13T12:04:27+02:00"), 'week')).to.eql(new Date("2023-05-14T23:59:59.999Z"));
+        });
+
+        it ('[week] should correctly set to end of week utc with monday as first day of the week when already on that day', () => {
+            expect(endOfUTC(new Date("2023-05-14T12:04:27+02:00"), 'week')).to.eql(new Date("2023-05-14T23:59:59.999Z"));
+        });
+
+        it ('[week] should correctly set to end of week utc with monday as first day of the week when end of week is in different month', () => {
+            expect(endOfUTC(new Date("2023-02-27T12:04:27+02:00"), 'week')).to.eql(new Date("2023-03-05T23:59:59.999Z"));
+        });
+
+        it ('[week] should correctly set to end of week utc with monday as first day of the week when end of week is in different year', () => {
+            expect(endOfUTC(new Date("2022-12-29T12:04:27+02:00"), 'week')).to.eql(new Date("2023-01-01T23:59:59.999Z"));
+        });
+
+        it ('[week] should be blazing fast in its conversion (1.000.000 benchmark < 750ms)', () => {
+            let start_time = getTime();
+            for (let i = 0; i < 1000000; i++) {
+                endOfUTC(new Date("2023-05-04T12:04:27+02:00"), 'week');
+            }
+            expect(getTime() - start_time).to.be.lt(750);
+        });
+
+        it ('[week_sun] should correctly set to end of week utc with sunday as first day of the week', () => {
+            expect(endOfUTC(new Date("2023-05-04T12:04:27+02:00"), 'week_sun')).to.eql(new Date("2023-05-06T23:59:59.999Z"));
+            expect(endOfUTC(new Date("2023-05-12T12:04:27+02:00"), 'week_sun')).to.eql(new Date("2023-05-13T23:59:59.999Z"));
+        });
+
+        it ('[week_sun] should correctly set to end of week utc with sunday as first day of the week when already on that day', () => {
+            expect(endOfUTC(new Date("2023-05-06T12:04:27+02:00"), 'week_sun')).to.eql(new Date("2023-05-06T23:59:59.999Z"));
+        });
+
+        it ('[week_sun] should correctly set to end of week utc with sunday as first day of the week when end of week is in different month', () => {
+            expect(endOfUTC(new Date("2023-03-29T12:04:27+02:00"), 'week_sun')).to.eql(new Date("2023-04-01T23:59:59.999Z"));
+        });
+
+        it ('[week_sun] should correctly set to end of week utc with sunday as first day of the week when end of week is in different year', () => {
+            expect(endOfUTC(new Date("2021-12-28T12:04:27+02:00"), 'week_sun')).to.eql(new Date("2022-01-01T23:59:59.999Z"));
+        });
+
+        it ('[week_sun] should be blazing fast in its conversion (1.000.000 benchmark < 750ms)', () => {
+            let start_time = getTime();
+            for (let i = 0; i < 1000000; i++) {
+                endOfUTC(new Date("2023-05-04T12:04:27+02:00"), 'week_sun');
+            }
+            expect(getTime() - start_time).to.be.lt(750);
+        });
+
+        it ('[day] should correctly set to end of day utc', () => {
+            expect(endOfUTC(new Date("2023-05-04T12:04:27+02:00"), 'day')).to.eql(new Date("2023-05-04T23:59:59.999Z"));
+        });
+
+        it ('[day] should be blazing fast in its conversion (1.000.000 benchmark < 750ms)', () => {
+            let start_time = getTime();
+            for (let i = 0; i < 1000000; i++) {
+                endOfUTC(new Date("2023-05-01T12:04:27+02:00"), 'day');
+            }
+            expect(getTime() - start_time).to.be.lt(750);
+        });
+
+        it ('[hour] should correctly set to end of hour utc', () => {
+            expect(endOfUTC(new Date("2023-05-04T12:04:27+02:00"), 'hour')).to.eql(new Date("2023-05-04T10:59:59.999Z"));
+        });
+
+        it ('[hour] should be blazing fast in its conversion (1.000.000 benchmark < 750ms)', () => {
+            let start_time = getTime();
+            for (let i = 0; i < 1000000; i++) {
+                endOfUTC(new Date("2023-05-01T12:04:27+02:00"), 'hour');
+            }
+            expect(getTime() - start_time).to.be.lt(750);
+        });
+
+        it ('[minute] should correctly set to end of minute utc', () => {
+            expect(endOfUTC(new Date("2023-05-04T12:04:27+02:00"), 'minute')).to.eql(new Date("2023-05-04T10:04:59.999Z"));
+        });
+
+        it ('[minute] should be blazing fast in its conversion (1.000.000 benchmark < 750ms)', () => {
+            let start_time = getTime();
+            for (let i = 0; i < 1000000; i++) {
+                endOfUTC(new Date("2023-05-01T12:04:27+02:00"), 'minute');
+            }
+            expect(getTime() - start_time).to.be.lt(750);
+        });
+
+        it ('[second] should correctly set to end of second utc', () => {
+            expect(endOfUTC(new Date("2023-05-04T12:04:27.043+02:00"), 'second')).to.eql(new Date("2023-05-04T10:04:27.999Z"));
+        });
+
+        it ('[second] should be blazing fast in its conversion (1.000.000 benchmark < 750ms)', () => {
+            let start_time = getTime();
+            for (let i = 0; i < 1000000; i++) {
+                endOfUTC(new Date("2023-05-01T12:04:27+02:00"), 'second');
+            }
+            expect(getTime() - start_time).to.be.lt(750);
+        });
+
+        it ('should return original date when passed a non-recognized key', () => {
+            expect(endOfUTC(new Date("2023-05-01T12:04:27+02:00"), 'foobar')).to.eql(new Date("2023-05-01T12:04:27+02:00"));
         });
     });